move Redirects to variables
